+++ Correlator/syslog.py
@@ -2,8 +2,10 @@ import iso8601
 import logging
 import re
 import socket
+
 from dataclasses import dataclass
 from mako.template import Template
+from time import sleep
 from typing import List, BinaryIO, Callable
 
 from Correlator.event import EventProcessor, ErrorEvent, AuditEvent
@@ -40,28 +42,36 @@ class SyslogServer:
 
     """
     def __init__(self, modules: List[Module], processor: EventProcessor,
-                 discovery_method: Callable[[RawSyslogRecord], bytes] = None):
+                 discovery_method: Callable[[RawSyslogRecord], bytes] = None,
+                 record_filter=None):
 
         self.modules = modules
         self.processor = processor
         self.buffer_size: int = DEFAULT_BUFFER_SIZE
         self.syslog_trailer = None
         self.trailer_discovery_method = discovery_method
+        self.bind_retry = 1
+        self.record_num = 0
+        self.record_filter = record_filter
+        self.output_file = None
 
-    def from_file(self, file_obj: BinaryIO):
+    def from_file(self, file_obj: BinaryIO, output_file: BinaryIO = None):
         """ Processes saved syslog data from a file
 
         :param file_obj: File object of a readable binary file
+        :type file_obj: typing.BinaryIO
+        :param output_file: File object of a writable binary file
         :type file_obj: typing.BinaryIO
 
         """
 
         last = b''
+        self.output_file = output_file
         while True:
             block = file_obj.read(self.buffer_size)
 
-            # Find our trailer (syslog record seperator) if
-            # We haven't already.
+            # Determine the trailer (syslog record seperator) if
+            # we haven't already.
 
             if self.syslog_trailer is None:
                 self.syslog_trailer = self.discover_trailer(block)
@@ -97,23 +107,32 @@ class SyslogServer:
             host = socket.gethostname()
 
         server_socket = socket.socket()
-        server_socket.bind((host, port))
-        server_socket.listen(2)
+
+        while True:
+            try:
+                server_socket.bind((host, port))
+                break
+            except OSError as e:
+                if e.strerror.lower() == 'address already in use':
+                    log.error(f'Address {host}:{port} already in use. Will '
+                              f'retry in {self.bind_retry} second(s)')
+                    sleep(self.bind_retry)
+                else:
+                    raise
+
+        server_socket.listen(1)
+        log.info(f'Server listening on {host}:{port}')
         conn, address = server_socket.accept()
-
-        (remote_host, remote_port) = address
-
-        if remote_host == '192.168.1.3':
-            self.syslog_trailer = b'\n'
 
         last = b''
 
+        self.output_file = output_file
         while True:
             block = conn.recv(self.buffer_size)
 
             # write to capture file, if desired
-            if block and output_file is not None:
-                output_file.write(block)
+            # if block and output_file is not None:
+            #     output_file.write(block)
 
             # Find our trailer (syslog record seperator) if
             # We haven't already.
@@ -133,15 +152,16 @@ class SyslogServer:
 
     def discover_trailer(self, block):
         raw_block = SyslogRecord.decode_from_raw(block)
-        try:
-            trailer = self.trailer_discovery_method(raw_block)
-            if trailer:
-                log.debug(f'Trailer discovery method returned {repr(trailer)}'
-                          f'. Using it')
-                return trailer
-        except Exception as e:
-            log.error('Trailer discovery method raised exception')
-            log.exception(e)
+        if callable(self.trailer_discovery_method):
+            try:
+                trailer = self.trailer_discovery_method(raw_block)
+                if trailer:
+                    log.debug(f'Trailer discovery method returned '
+                              f'{repr(trailer)}. Using it')
+                    return trailer
+            except Exception as e:
+                log.error('Trailer discovery method raised exception')
+                log.exception(e)
 
         log.debug(f'Using default trailer of {repr(DEFAULT_SYSLOG_TRAILER)}')
         # Default trailer
@@ -158,6 +178,18 @@ class SyslogServer:
 
     def _process_record(self, data):
         record = SyslogRecord(data)
+
+        if self.output_file is not None:
+            # Write record to capture file
+            process_record = True
+            if self.record_filter is not None:
+                # Unless were supplied a record_filter, and this record is
+                # not to be written.
+                process_record = self.record_filter[self.record_num]
+            if process_record:
+                self.output_file.write(data + self.syslog_trailer)
+
+        self.record_num += 1
         if not record.error:
             for module in self.modules:
                 module.process_record(record)
@@ -190,10 +222,12 @@ class SyslogRecord:
 
     @staticmethod
     def decode_from_raw(block):
-        """Finds the first occurrence of structured data in a raw data block.
+        """Parses the data-only* portion of a syslog record in a raw data block.
 
-        Returns a dict representation of the structured data.
-        This is used only for syslog trailer discovery.
+        This is used to parse the first block to be used for syslog trailer
+        discovery.
+
+        * data-only means all properties up to but not including detail.
 
         """
 
@@ -242,10 +276,12 @@ class SyslogRecord:
 
         self.timestamp_str = self.m.group('timestamp_str')
         try:
-            self.timestamp = iso8601.parse_date(self.timestamp_str)
+            timestamp_tz = iso8601.parse_date(self.timestamp_str)
         except iso8601.ParseError:
             self.error = 'Cannot parse timestamp'
             return
+
+        self.timestamp = timestamp_tz.replace(tzinfo=None)
 
         self.priority = self.m.group('priority')
         self.hostname = self.m.group('hostname')
@@ -311,11 +347,12 @@ class SyslogRecord:
                         f'SD-DATA Key/Value {dataline} parse failed')
 
     def __repr__(self):
-        # why?
+        # todo: why?
         return f'{self.m.groupdict()} ({self.structured_data})'
 
     def __str__(self):
-        return f'{self.hostname} {self.appname} {self.proc_id} {self.detail}'
+        return (f'{self.timestamp.strftime("%Y-%m-%d %H:%M:%S")}: '
+                f'{self.hostname} {self.appname} {self.proc_id} {self.detail}')
 
     def __len__(self):
         return self.record_length